microservices_chatbot/ai/requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
astor==0.8.1
blinker==1.4
cached-property==1.5.2
click==8.1.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.46.1
h11==0.13.0
h2==4.1.0
h5py==3.6.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.21.6
opt-einsum==3.3.0
priority==2.0.0
protobuf==3.20.1
pydantic==1.9.0
pyhumps==3.7.1
quart==0.17.0
quart-schema==0.11.1
six==1.16.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
